+++ src/components/AnalysisResults.js
@@ -1,5 +1,5 @@
 // AnalysisResults.js
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import { handleFilter, resetFilters } from '../utils/Filters';