+++ CHANGELOG.md
@@ -47,3 +47,6 @@ Results.js:
 - Displays products and ingredients even if the user doesn't select any skin concerns
 - Added a loading spinner
 - Split SkinAnalysis.js into two files, analysis + results
+
+1/21/25 - Hannah
+- Automatically filters products based on routine steps chosen in skin quiz